refactor command handling: migrate args functionality to server package and remove deprecated args module
